up/down/top/bottom: Fix directions
The directions were reversed.
`up` and `top` should move upstack,
`down` and `bottom` should move downstack.

Upstack is defined as branches that are further away from the trunk.
Downstack is defined as branches that are closer to the trunk.
The direction here is further cemented by the "base" terminology
which refers to the parent branch--the 'down'-stack.