cmd up: Test CLI interaction
If we're going to do terminal UI, we should test it.
testscript's built-ins for this aren't great,
but we can hack one on with relative ease.

This creates a with-term wrapper command
that will exceute the provided command with a pty attached,
and a vt100 terminal emulator to view the result.

The result is that we get a second simple language
for interacting with the terminal application:

    snapshot [name]
      Print the current view to stdout,
      optionally with a title.
    await [txt]
      Wait for [txt] to appear on screen,
      or for screen to change since last snapshot.
    feed txt
      Feed txt into the terminal command.

We can tweak this language/incorporate it into testscript later
if this becomes too much.

This also found a bug in the implementation of the prompt:
it always selects the first branch.