{up,down,bottom,top}: Share checkout logic
Share checkout logic between movement comands
and always report when a branch needs to be restacked.
Also adds multi-head handling to 'up'.

Right now, this repeats some work
as we'll constantly re-open the repo and data store.
In a future change, we can add sharnig of internal state
between commands that rely on each other like this.