commit create, upstack restack: Hide annoying message
Commands that invoke branch restack
currently print an annoying "X does not need to be restacked" message
which is nearly always true for those commands.

Avoid printing that message with a small hack.

In a future change, we'll want to use a different mechanism
for commands to invoke each other,
that will provide better sharing of information between them.
We'll be able to drop that log message in that case.