Move "verify restacked" logic into Service, auto-fix state
Moves the logic to check if a branch is up-to-date with its base
into gs.Service, which makes sense as a place for that logic.

Changes the logic to also auto-fix the base branch hash stored in the
state if the user manually restacked out of band when we weren't
looking.