repo init: Generalize branch/trunk guessing logic
This commit introduces a new `Guesser` type that encapsulates the logic
for guessing the state of a repository during initialization.

It hides the prompt logic away from the internals
so we can implement non-interactive mode later.

This abstraction will probably be renamed in the future
if we add other re-usable UI bits to it.