flip flop on navigation yet again
up/down/top/bottom are branch scoped but also stack scoped
and it's annoying that there isn't a clear boundary.
This is the Nth time we've flip flopped on it.

Move them back to top level commands,
but add 'b' prefixed aliases to indicate branch-level-ness.

This may bear further reconsideration in the future.