fix(branch track): Don't auto-guess self, always reject cycles
'gs branch create' had a bug where,
when auto-guessing a base branch for a branch,
if the target branch was already tracked,
it would use self as the base branch.
Fixing that is easy: leave self out of consideration,
but it also exposed the fact that it's possible to create tracking
cycles.

This fixes the prior bug, and adds a check to 'branch track',
ensuring that adding that connection would not cause a branch cycle.