ui: Replace huh with custom widgets
Use of huh was always temporary. We want tighter control over the UI,
so we'll build our own widgets on bubbletea.

This replaces use of huh with custom bubbletea widgets
with an API similar to huh.

Chief among the new widgets is the select widget
which provides fuzzy search and selection of items,
and supports scrolling through a large list of items.

Resolves #88