Add CLI completion
This adds bash, zsh, and fish completion for the CLI.
The logic for this is basically a copy of how Kong's CLI parsre works.
There may be some edge cases (it needs unit tests)
but manual testing indicates that this works reasonably well.

This does not use kongplete largely because:

- it appears to be unmaintained
- it does not expose the completed argument list
  (which is necessary for handling shorthand commands)
- I'm not entirely sure it implements the same parsing logic as Kong

We should consider upstreaming this logic into Kong in the future;
perhaps the duplication can be avoided.

Resolve #32