restack: Use --is-ancestor, share logic with VerifyRestacked
To check if a branch is already restacked, use --is-ancestor.
It's better than checking if the merge-base of the two branches
is the base branch.

Also, instead of duplicating the logic in Restack,
share it with VerifyRestacked.
For efficiency, pass information back from VerifyRestacked to Restack
through the BranchNeedsRestackError type.