fix(checkout): don't report untracked for absent
If you run 'branch checkout foo' for a foo
that does not exist, branch checkout would report
that the branch is untracked
and that the user should run 'gs branch track foo',
before reporting that the branch does not exist.

The reason for this was that LookupBranch did care about non-existence
and returned all branches as untracked when they didn't exist.
The handling for this is a bit messy, but basically we have to handle
four different scenarios when looking up a branch.