git/rebase: Detect interrupts and provide a hook
If a rebase operation fails because it was interrupted,
whether due to conflict or because of a user instruction,
we don't want to fail the entire git-spice operation.

To fix this, we need a means to handle the rebase failure
with custom logic.

This adds support for a function to be run when a rebase is interrupted.
The function gets information about the state of the rebase,
extracted from rebase's internal state.

This isn't used yet by git-spice commands that perform rebases.