Adjust import path, move command to cmd/gs
Change the import path of the project to use "git-stack",
same as the repository name.
We still want to use the command name "gs", so move the actual CLI
to a cmd/gs subdirectory.

So installation from source will now be

    go install go.abhg.dev/git-spice/cmd/gs

=== BREAKING ===

This also adjusts the ref name under which data is stored
from refs/gs/data to refs/spice/data.

Migration steps for alpha users:
Run the following command when you pull the new binary.

	git update-ref refs/spice/data $(git rev-parse refs/gs/data)
	git update-ref -d refs/gs/data