rebase continue: recover from rebase conflicts
This adds support for continuing or aborting an interrupted rebase-ish
operation with two new commands:

    gs rebase continue  # alias: gs rbc
    gs rebase abort     # alias: gs rba

The general idea is that when a rebase-ish operation is interrupted,
we'll record the command that should run after the rebase is resolved,
and the user can run 'gs rebase continue' to continue the operation.

This works for all commands that call `git rebase` under the hood,
including convenience commands like `gs commit create`.

Resolves #42