branch submit: Fix PR title and body for multiple commits
git rev-list returns the commit list in reverse order
(newest commit first).
This is right for most of our needs,
except when generating the default PR title and body,
we need to consider it in reverse:
earliest commit should be the PR title,
and the commit bodies should be in the PR description
in chronological order.

Resolves #69