restack: Don't use fork point if base hash is up to date
When restacking, we run the equivalent of the following command:

    git rebase --onto [base] [upstream] [current]

Where,

- [base]: name of the base branch we're stacking on top of
- [upstream]: commit to start rebasing from
- [current]: branch we're rebasing

Commits that match `git rev-list [current] --not [upstream]`
are included in the rebase.

We track the hash of the base branch as the user uses git-spice
so that we know the best value of [upstream] to use when restacking.

However, our restacking logic was always overriding that
with the fork point between [base] and [current] if available.
This seems unnecessary and possibly faulty
given that our base branch information is more specific.

In this change, we'll only use the fork point as a fallback
if the recorded base hash is not an ancestor of the current branch.