Fix 'branch onto', 'upstack onto' discrepancy
The name 'branch onto' suggests that it extracts just the given branch,
while the name 'upstack onto' suggests that it extracts the branch
and its entire upstack.

Prior behavior of 'branch onto' was what we want 'upstack onto' to do,
minus the 'upstack restack' at the end.

This change renames 'branch onto' to 'upstack onto',
and adds a new 'branch onto' command that extracts just that branch.
It specifically removes that change from the stack that it was extracted
from. ('branch delete' for #57 probably wants to do something similar.)

Resolves #89