branch delete: Nuke deleted branch from stack
Previously, when a branch was deleted, we stopped tracking it,
but its changes were still present in the stack.

This fixes this incorrect behavior:
now, the deleted branch's upstack will be rebased onto its base,
deleting its changes from the stack.

The old behavior is still available with `branch untrack`.