rebase {continue, abort}: Support multiple continuations
It's possible to have multiple divergent rebase operatons
step from the same re-entrant command.
In particular, having a single continuation breaks for `branch onto`,
which needs to run `upstack onto` on the entire upstack
before updating the branch's own state.

To fix this, we need to use a queue of continuation commands.