repo sync: Don't leave on the wrong branch
'repo sync' deletes merged branches and restacks the upstacks.
Previously, there was a bug in `branch delete` that left us
on the wrong branch afterwards (the upstack).
This was fixed in #137, but the issue was discovered in #130.

Adds a test case to verify the test case for the bug in #130.

Resolves #130