refactor: Abstract GitHub client away
Adds a layer between GitHub client and business logic currently referred
to as "forge" with the idea that we may support non-GitHub code forges
in the future.

The abstraction layer refers to PRs as Changes to further disconnect
from GitHub-specific terminology.

This also fixes a previously known bug where GITHUB_GIT_URL
was not actually accounted for when matching the remote URL.
In this change, that's renamed to GITHUB_URL, and is actually used
when determining the GitHub repository.

The internal/gh package has been merged into this new package
while we're at it.