WIP: GitHub: Use GraphQL API
Some functionality is gated behdind the GraphQL API
(e.g. draft status #84, PR templates #87),
so we have to use this anyway.

Instead of bundling both and significantly inflating the binary size,
just switch to GraphQL for all operations.

On the bright side, this shaves off 4MB from the binary size.

- [ ] Figure out how to test end-to-end flow since ShamHub is broken
      now. Might have to test against the Forge interface instead.
- [ ] In a follow-up, persist GraphQL IDs for repo and PRs in the store
      to avoid looking them up every time.