rebase continue: Recover from 'rebase --abort'
If a user runs into a conflict with a `gs` operation,
and then they run `git rebase --abort` instead of a `gs rebase` command,
we'll leave behind a rebase continuation in the data store
which will then be called by the next `gs rebase continue` command.
This is undesirable and can lead to unexpected behavior.

Fix this by clearing the rebase state at the start of a rescue.

For this to work correctly with `gs rebase continue`,
the continue command has to restore the remaining continuations
if a prior continuation fails
because otherwise the continuation will clear the other ones.