branch submit: Respect PR templates
Picks up the first available PR template, if any
and fills the default PR body with it (along with the commit messages).
Allows for users to fill in details, and adjust the PR message
before submission.

Caveat: This will perform a PR template look up on every PR submission.
For a 'stack submit', that means N lookups.
We can probably optimize this by caching the PR template
between submissions in the same command invocation,
or by generating some sort of hash key based on the .github/ directory
and store it in the data store.

Resolves #87