add 'branch split' command
Adds a 'gs branch split' (gs bsp) command
that allows introducing a new branch on an existing commit.

In non-interactive mode, commits to split on are specified explicitly
with `--at`.

In interactiev mode, the user is presented with a list of commits
and they mark the ones they want to split on.
Names are then requested for the new branches.

Resolves #33