Add auth commands, use Device Flow or PAT
Finally drops the hack of shelling out to `gs auth login`
in favor of explicit authentication.
Multiple authentication methods are supported:

- OAuth with repo or public_repo scope
- GitHub App
- Personal Access Token

Setting `GITHUB_TOKEN` supersedes other login methods,
which allows for scripting in environments where prompting
for login choices isn't an option.

After authentication, the token is stored in a system-specific
secure storage.
Because this limits the number of supported systems,
in the future we can add a fallback plain text storage.

This PR was made with GitHub App authentication.
Note that at this time, the GitHub App is private
and cannot be installed on anyone else's account.
Beta users should use OAuth or PAT.

Resolves #22