branch submit: Recover from failed pushes
If a 'branch submit' fails to actually submit
(due to a failing pre-push hook, for example)
we can lose whatever PR metadata the user has already filled in.
This can be annoying if the PR body diverges significantly
from the commit body because some amount of work is lost.

As a UX improvement here, we'll record the title/body
after the user fills the form in case the submission fails.
We'll recover this saved information when the user re-submits,
prompting them to let them use the recovered information,
and edit it if necessary.

Resolves #221