Add 'commit split' command
Adds a new 'gs commit split' (gs csp) operation
that will split the top commit into two commits,
with the user selecting which hunks to move to
the new commit.

The new commit will be placed below the current
commit, and branches upstack will be restacked
as needed.

This operation may be extended in the future
for splits further down the commit history.