branch submit: Fix directory+ref conflict
'branch submit' doesn't work if we have a directory (or file)
in the root of the repository with the same name as the branch.
It fails because we use git-rev-list to resolve information about the
branch and commits in it, and git rev-list needs you to disambiguate
by using `--` to separate the branch name from the path.

```
> gs branch submit --fill
FTL gs: list commits: rev-list: exit status 128
stderr:
fatal: ambiguous argument 'doc': both revision and filename
Use '--' to separate paths from revisions, like this:
'git <command> [<revision>...] -- [<file>...]'
```

To fix this, add '--' to git rev-list invocations
to disambiguate the branch name from the path.