test: Add fixturetest to generate semi-random string
The integration tests for GitHub use a repeated pattern
where they generate a random string if run with `-update`,
but store those in files for later runs for deterministic behavior.

As we'll be needing more of this, this extracts that functionality
into a separate fixturetest package.
This allows for a pluggable `func() T` to generate the value,
and stores it in a file as JSON for later runs.

[skip changelog]: Not user facing