all/submit: Share state between submit commands
Add the ability to share some amount of state between the various
submit commands so we're not repeating initialization work.

The immediate effect of this is that we can share
information about the remote repository between submits,
which avoids one "get repository GraphQL ID" call per branch.

Long-term, we should just store the GraphQL ID in the local store
as it's not going to change.