submit: Post comments with a stack visualization
For all CR submission commands,
track which branches are being submitted.

For all such branches, add a comment with a visualization of the stack
marking the position of that CR in the stack.

The visualization does not include the entire stack:
for each CR, it only includes changes that are downstack from it,
the CR itself, and the changes that are upstack from it.
It does not include siblings or cousins.

Resolves #142