all: Normalize terminology to "CR"
We've been using Pull Request and Change Request in messaging
interchangeably.
Originally, most code was written with Pull Request as the tool was
coupled to GitHub.
However, with the refactor to extract the Forge type,
we can now support other code forges, so don't need that coupling.
Since then, messaging uses "CR" or Change Request.

Take a pass over any user-facing messages that are referencing PRs
instead of CRs and fix them.