fix: Use 'git var GIT_EDITOR' for PR body
We were using `$EDITOR` to fill in the PR body.
This is not correct.
We should use the output of 'git var GIT_EDITOR'.
The Git editor (and `$EDITOR`) are allowed to be shell commands,
not just executable names. Handle that correctly too.

Lastly, re-use the same logic for stack edit and downstack edit.

Resolves #274