forge/github: Determine API URL from base URL if not set
For GHES, usually the API URL is /api under the base URL.
So if the API URL is not set, and base URL is not github.com,
we can assume the API URL is $baseURL/api.

Users can still override this with the `spice.forge.github.apiUrl`
or `GITHUB_API_URL` environment variable.

Follow up to https://github.com/abhinav/git-spice/pull/330#issuecomment-2277077683