github: Fix outdated PR templates with lower-case names
We use the contents of possible PR template sites
as the hash key for caching the PR templates we get from the remote.
This was only looking at PULL_REQUEST_TEMPLATE.md names,
even though pull_request_template.md is valid and accepted.

This change ensures that lower-cased versions are also considered
when calculating the hash key.

The contract of ChangeTemplatePaths has been adjusted
to state that all paths are case-sensitive, not the other way around,
since we don't have a guarantee that all forges will treat these
as case-insensitive.

Resolves #369