branch submit: Handle CR already closed or merged
If a submitted CR is already closed or merged server-side,
pretend that the CR does not exist and create a new CR instead.

This is a more user-friendly behavior
than permanently tying the branch name to that CR
(or until the user runs untrack-retrack).

In the future, we can offer to re-open closed CRs with a prompt too.

Resolves #374