repo sync: Don't complain about deleted remote branches
Right now, if you run `git pull` or `git fetch` with `--prune`
before calling `repo sync`,
it will complain that the remote tracking branch is missing.
It's only an error, not a warning, but it can still be confusing.

This makes `repo sync` check if the remote tracking branch exists
before trying to delete it,
making it a no-op if the branch has already been deleted.