repo sync: Gracefully handle unsupported forges
When `repo sync` is run on a repository with a remote
that is not a known forge (e.g. GitLab),
instead of failing with an error because we can't detect merged PRs,
just log a warning and skip the branch deletion step.

This is only a partial fix for #391.
It'll still leave the merged branches around,
and worse, the upstack branches will still point to merged branches
as their bases.
Since we can easily detect merged branches that used a merge commit,
we will do that instead of a warning in a follow-up PR.