repo sync: Delete merged branches for unknown forges
Follow up to #404.
Previously, we just skipped branch deletion for unknown forges.
This degrades more gracefully by doing a best-effort check
for merged branches by looking at what's reachable from the trunk.

This will not handle squashes or rebases,
but it's good enough in lieu of official support for the forge.

Resolves #391