WIP: state: formalize BranchGraph
Instead of building and discarding the branch graph for UpdateBranch,
make BranchGraph the primary interface for managing state.

Branch state will be loaded lazily as needed,
and changes will be accumulated in the BranchGraphTx
until Commit is called.

Right now, the store's branch methods are just wrapping the graph.
In a future change, those will be deleted in favor of the graph.

[skip changelog]: No user facing changes.J ust a better version of #419.