forge/github: Surface GraphQL API error codes
shurcooL/githubv4 hides the error codes in the returned errors.
There's an open issue to resolve this but it hasn't seen progress
in years.

As a workaround, we'll interpret the GraphQL errors ourselves
at the HTTP transport layer -- at least for now.
In the future, we should switch to a different GraphQL client library.

Closes #437