submit: Generate unique remote name if branch name is taken
When submitting a branch `$name`,
it's possible that someone else has already taken that name.
For example, the local branch name might be something generic
like 'feature', and someone else may already have taken that name.

To handle this case, gs will now query the remote with `git ls-remote`,
and generate a unique name for the branch if the original name is taken.
New names start at `$name-2`, and keep incrementing the number
until a free name is found.

Making this change results in a couple other changes:

- If the branch has an upstream associated with it,
  e.g. with `git push -u origin foo:bar`, we'll prefer that name
  since the user has already pushed the branch and chosen it.
  This makes sense: I pushed the branch, and associated the upstream
  with the branch, so that's what should be used.
- When matching externally submitted CRs to the branch,
  just matching on the branch name is no longer sufficient.
  If we're handling conflicting names, we have to account for the case
  where someone else submitted a CR with the conflicting branch name.
  Therefore, CR matching now also requires that the HEADs match.

TODO:

- [ ] When ignoring CR, print a message suggesting they should cancel
      this operation and 'git pull' that branch if they're sure it's that CR.
- [ ] When using the branch name from `@{upstream}`,
      print a message informing the user of this.

Resolves #408