refac: record hostname on token for CLI, extract githuCLI abstraction
I like to avoid having multiple places invoke the same CLI separately
with similar-ish handling of output/etc. In such a case, it's useful
to put the os/exec's behind an abstraction of some kind.
In this case, we have two: Status and Token.
This extracts the CLI interfacing into this abstraction,
and installs an interface based on it for the authenticator.

Additionally, _when_ we get the token is changed:
instead of LoadAuthenticationToken which should just save and load
the AuthenticationToken type from the stash,
we now get the token when we need it: when creating the client.

Otherwise, the AuthenticationToken was in this weird state where when it
was AuthTypeGitLabCLI, it was expected to have an empty AccessToken, but
then it had a token when calling LoadAuthenticationToken.

With this, AuthenticationToken now always have one of the following:

- AuthTypePAT: AccessToken is set
- AuthTypeOAuth2: AccessToken is set
- AuthTypeEnvironmentVariable: AccessToken is set
- AuthTypeGitLabCLI: AccessToken is empty, Hostname is set