rebase continue: Add --no-edit flag to skip opening an editor
Git's `rebase --continue` always opens an editor
to edit the commit message after resoling the conflict.
There are cases where this is not necessary.

This adds a `--no-edit` flag to `gs rebase continue`
to skip opening an editor.

`rebase` does not have a `--no-edit` flag to skip the editor
but we can get the same effect by setting the editor to `true`
for that command.

Resolves #503