submit: Include merged CRs in navigation comments (#508)
This change adds merged CRs to navigation comments.
When a CR is merged, its upstacks will still report it in their nav comment
even when they're rebased and resubmitted.


To do that git-spice now keeps track of merged PRs in the per-branch state
of the bottom branch in each stack.
This then gets bubbled up to all branches directly above it when that branch
is merged.

The state is currently propagated from `repo sync`.
Two special-cases are added:

- `stack edit` and `downstack edit` will transfer
  the merged history  to the new bottom-most branch
- `branch create --below` from the bottom-most branch
  will transfer merge history to the new bottom-most branch

In the future, we may need to introduce a shared space
for tracking branch and CR history instead of in dependents.
(e.g. a place where all branches, current and former are tracked,
maybe with some garbage collection.)


Resolves #331

---------

Co-authored-by: Abhinav Gupta <mail@abhinavg.net>