test(repo sync): navigation history for unsubmitted branches
Add a test for how merge history propagation behaves
for upstack branches that are unsubmitted.

Suppose I have `feat1 -> feat2`:

- feat1 is ready and I submit a CR
- feat2 is not yet ready for review, so it is unsubmitted
- while I'm still working on feat2, feat1 gets approved and I merge it
- I then submit feat2

Right now--by accident--it works so that feat2 receives the history:
it will show feat1 as its merged downstack even though feat1 was merged
by the time feat2 was restacked on main and submitted.

My intuition is that this should NOT be supported,
but it's also coming mostly for "free".

I'm opening up this PR and test case to discuss this behavior
with anyone who cares.