fix(gitlab): Actually support GITLAB_TOKEN (#613)
The GitLab forge was rejecting use of the GITLAB_TOKEN env var
for authentication despite advertising support for it.

The issue was that the client constructor was exploding
when the auth type was `AuthTypeEnvironmentVariable`
because it was not in the list of known auth types.

Fix this by adding a case for it, and add tests to verify.

Resolves #612