ci: Run tests with race detection
PR #636 introduced parallel workers for log data collection.
We may add more parallelism in the future.

Race detection was previously disabled to keep test execution time
low back when it was sub 30 seconds.
With Windows tests, it takes 5 minutes anyway,
so there's no reason to keep it disabled for Linux.