state: detect upstream ref deletion (#662)
If an upstream branch reference is deleted manually
(e.g. with git push --delete, or by deleting the remote branch),
then the state's local knowledge of the upstream branch name is stale
and should be deleted.

This does not solve, but helps with issues like #655,
as the user won't have to untrack the branch
to delete stale information.