fix(branch create): prefix in branch name conflict resolution (#676)
When 'branch create' generates a branch name based on the commit,
if the branch name already exists, it appends a number to the name.
Right now, we add the prefix to the generated name afterwards,
which means that there's a chance that the final name
will conflict with an existing branch.

This change fixes this by ensuring that the prefix is taken into account
when generating the branch name.