branch create: reject untracked base branch earlier (#708)
branch create rejects untracked base branches,
but only after the user has written a commit message.
This is a form of lost work, especially if the commit message is long.

This changes branch create to reject untracked base branches earlier,
so that the user does not have to write a commit message first.

Resolves #652