submit: add 'spice.submit.draft' to change --draft/--no-draft default (#721)
Add a spice.submit.draft configuration option
that will control the default value of --draft/--no-draft
when creating new CRs with `gs branch submit` and friends.

In interactive mode, this merely controls the default value
of the prompt that asks for the draft status.
(So a user can just press Enter to accept the default.)

In non-interactive mode, this sets the value that will be used
if the user does not explicitly pass --draft or --no-draft.

Tests include all combinations of:

    { interactive, non-interactive } x { --draft, --no-draft, no flag }

Refs #568