fix: concurrent Git storage usage, bad base hash after manual rebase (#736)
This PR fixes a couple issues discovered in trying to fix #731.

First, there's a data race in the Git backend:
when multiple processes try to write to it,
the `update ref` operation can fail.
We accounted for this and retry, but with enough updates,
the operation can actually fail.

To fix this, add locking around the Git backend's read/write operations,
ensuring that only one process can write at a time.
The included test tries to run a bunch of operations on it concurrently,
and fails without the fix.

Secondly, we discovered that in a couple places, including `log long`,
we did `LookupBranch`/`LoadBranches` before calling `VerifyRestacked`.
This means we loaded the branch information (including the base hash)
and then called `VerifyRestacked`, which, despite the read-only-looking
name
may change the branch's base hash if the branch was manually restacked
by the user (using `git rebase` instead of `gs branch restack`).
This would result in the operation using the wrong base hash.

The following commands were affected:

- `log long` would show the wrong list of commits
  for the first run after a manual rebase.
- `branch squash` would use the wrong base hash when squashing the
commits,
  resulting in an extraneous no-op commit.
- `branch submit` was not negatively affected, but it was updated
nonetheless

(Test cases included for `log long` and `branch squash`.)

Resolves #731